+++ lib/put-item.js
@@ -1,7 +1,7 @@
 const ErrorHTTP = require('mapbox-error').ErrorHTTP;
 
 const _ = require('lodash');
-const db = require('./db');
+const db = require('../database/db');
 const TaskItems = db.TaskItems;
 // const updateStats = require('./update-stats');
