+++ lib/update-stats.js
@@ -1,4 +1,4 @@
-const TaskUserStats = require('./db').TaskUserStats;
+const TaskUserStats = require('../database/db').TaskUserStats;
 const constants = require('./constants');
 
 /**
