+++ routes/item-comments.js
@@ -1,0 +1,37 @@
+module.exports = {
+  getItemComments: getItemComments,
+  getItemComment: getItemComment,
+  createItemComment: createItemComment,
+  updateItemComment: updateItemComment,
+  deleteItemComment: deleteItemComment
+};
+
+/**
+ */
+function getItemComments(req, res, next) {
+  return next();
+}
+
+/**
+ */
+function getItemComment(req, res, next) {
+  return next();
+}
+
+/**
+ */
+function createItemComment(req, res, next) {
+  return next();
+}
+
+/**
+ */
+function updateItemComment(req, res, next) {
+  return next();
+}
+
+/**
+ */
+function deleteItemComment(req, res, next) {
+  return next();
+}
