+++ routes/item-tags.js
@@ -1,0 +1,37 @@
+module.exports = {
+  getItemTags: getItemTags,
+  getItemTag: getItemTag,
+  createItemTag: createItemTag,
+  updateItemTag: updateItemTag,
+  deleteItemTag: deleteItemTag
+};
+
+/**
+ */
+function getItemTags(req, res, next) {
+  return next();
+}
+
+/**
+ */
+function getItemTag(req, res, next) {
+  return next();
+}
+
+/**
+ */
+function createItemTag(req, res, next) {
+  return next();
+}
+
+/**
+ */
+function updateItemTag(req, res, next) {
+  return next();
+}
+
+/**
+ */
+function deleteItemTag(req, res, next) {
+  return next();
+}
