+++ test/status.test.js
@@ -16,7 +16,7 @@ test('GET / - check that status is OK when db connection works', [], function(
 // Directly use tape to mock db
 tape('GET / - the status is NOTOK when db connection fails', function(assert) {
   const server = proxyquire('../lib/server', {
-    './db': {
+    '../database/db': {
       // make the mock db throw unconditional error
       authenticate: () => {
         return Promise.reject(new Error('error'));
